+++ homematicip/aio/device.py
@@ -461,3 +461,6 @@ class AsyncGarageDoorModuleTormatic(GarageDoorModuleTormatic, AsyncDevice):
 
     async def send_door_command(self, doorCommand=DoorCommand.STOP):
         return await self._connection.api_call(*super().send_door_command(doorCommand))
+
+class AsyncPluggableMainsFailureSurveillance(PluggableMainsFailureSurveillance,AsyncDevice):
+    """ [HMIP-PMFS] (Plugable Power Supply Monitoring) """