+++ homematicip/device.py
@@ -1775,3 +1775,21 @@ class GarageDoorModuleTormatic(Device):
     def send_door_command(self, doorCommand=DoorCommand.STOP):
         data = {"channelIndex": 1, "deviceId": self.id, "doorCommand": doorCommand}
         return self._restCall("device/control/sendDoorCommand", json.dumps(data))
+
+class PluggableMainsFailureSurveillance(Device):
+    """ [HMIP-PMFS] (Plugable Power Supply Monitoring) """
+
+    def __init__(self, connection):
+        super().__init__(connection)
+        self.powerMainsFailure = False
+        self.genericAlarmSignal = AlarmSignalType.NO_ALARM
+
+    def from_json(self, js):
+        super().from_json(js)
+        c = get_functional_channel("MAINS_FAILURE_CHANNEL", js)
+        if c:
+            self.set_attr_from_dict("powerMainsFailure",c)
+            self.set_attr_from_dict("genericAlarmSignal",c,AlarmSignalType)
+
+    def __str__(self):
+        return f"{super().__str__()} powerMainsFailure({self.powerMainsFailure}) genericAlarmSignal({self.genericAlarmSignal})"