+++ tests/test_devices.py
@@ -1192,12 +1192,25 @@ def test_door_sensor_tm(fake_home: Home):
         d = fake_home.search_device_by_id("3014F0000000000000FAF9B4")
 
         assert d.doorState == DoorState.CLOSED
-        assert d.on == False
-        assert d.processing == False
-        assert d.ventilationPositionSupported == True
+        assert d.on is False
+        assert d.processing is False
+        assert d.ventilationPositionSupported is True
 
         assert str(d) == (
             "HmIP-MOD-TM Garage Door Module lowbat(None) unreach(False) rssiDeviceValue(-52) "
             "rssiPeerValue(-54) configPending(False) dutyCycle(False) doorState(CLOSED) "
             "on(False) processing(False) ventilationPositionSupported(True)"
+        )
+
+def test_pluggable_mains_failure(fake_home:Home):
+    with no_ssl_verification():
+        d = fake_home.search_device_by_id("3014F7110000000000ABCD50")
+
+        assert d.powerMainsFailure is False
+        assert d.genericAlarmSignal is AlarmSignalType.FULL_ALARM
+
+        assert str(d) == (
+            "HmIP-PMFS Netzausfallüberwachung lowbat(None) unreach(False) rssiDeviceValue(-58) "
+            "rssiPeerValue(None) configPending(False) dutyCycle(False) powerMainsFailure(False) "
+            "genericAlarmSignal(FULL_ALARM)"
         )